Mavenize Barcode_Codec
